Bug: undo wasn't working on containers
Apparently every assignment to tag.value adds to the undo stack.
Essentially, Undo triggered a SelectionChanged event (before Undo had a
chance to work... no idea why this is the order of events). After
SelectionChanged, we asked for our options for the given selection.
Since Undo hasn't happened, we detect that it's a BlokContainer and call
BlokAdapter.getBlokContainer.

getBlokContainer always calls setSavedProperty, which ALWAYS assigned to
tag.value. This added to the undo stack before Undo had a chance to do
what the user actually wanted. Thus, Undo undid the tag.value assignment
that setSavedProperty performed!

To the user, it looks like Undo did nothing. And it's an infinite loop.
The fix is to only assign to tag.value if we need to.