Plugin only emits events
Plugin emits events for selection changed and pre-undo. Removed
functionality to listen to events. Trying to hook up the silent
transform was a PITA. We'd have to recover the pageItem reference from
across the ExtendScript boundary. On top of that, I had no success with
the Undo suite and setting silent. It's super confusion. Might as well
change our strategy.