Bug: stretching a parent BlokContainer incorrectly stretches a child BlokContainer
Nested BlokContainers can't be stretched the way normal Bloks can. The
scenario of a stretch parent BlokContainer that contains another stretch
child BlokContainer does not cause the child to stretch to the parent,
since the child's bounds are fixed. You would have to manually adjust
the bounds. Maybe we'll re-examine this scenario if we get feedback.