Add a Layout button
For manually triggering a layout when we can't be smart enough to do it
for the user.

A scenario: User has nested BlokGroups. User is editing one in isolation
mode, resizing contents. User exits isolation mode. At this point we
have no selection to work with. Even if we tried to guess the selection
is the first available BlokGroup, it's totally possible the bounds of
that BlokGroup haven't changed and thus aren't dirty. We'd have to start
caching the x,y coords of all Bloks, then check not only the current
Blok for being dirty, but also it's children (honestly, all the way down
the tree). Maybe this isn't too expensive, but man it would be a
headache to write. Let's just give the user the option to fix us.