Hook up cascading layout via Symbols
This one was tricky since Illustrator's scripting API doesn't expose a
way to see if a SymbolItem has an additional transform applied overtop
of the underlying Symbol, which makes it difficult to figure out whether
fixedWidth or fixedHeight needs to be updated for that Blok that is also
a SymbolItem.

It appears that there is a native API for getting that information, but
of course going down to that level is a pain.

So the easier fix is to detect when we're in "Symbol Editing Mode",
figure out what the dims are of the raw Symbol, then find all
SymbolItems that are also Bloks that use that Symbol, then update their
fixedWidth/fixedHeight.

Illustrator will refuse to run the transform() while we're editing a
symbol, so we cache the Bloks that needs to be invalidated (our first
global variable!), and invalidate them when we get a chance (usually you
can just exit Symbol Editing Mode and it'll just happen).

The caveats: if you're editing a Symbol with multiple sub layers or a
grouped object, you must select the root item for the change to be seen.
Also, if you transform a SymbolItem, we'll just blow away that
transformation in terms of width/height.