Bug: BlokContainers with type in them and Align to Pixel Grid refuse to layout
1. Open a document with 2 PathItems that are on the pixel grid AND have
the "Align to Pixel Grid" setting checked
2. Insert some type in between the 2 Path items
3. Create a horizontal BlokContainer

ACTUAL: BlokContainer would refuse, since checkForRelayout() would find
that width is invalid and app.undo()
EXPECTED: Well, we can't have that can we?

Illustrator will never let you check "Align to Pixel Grid" for type
(kind of makes sense), so it's almost always off. However, strangely
type can be part of a GroupItem that does have "Align to Pixel Grid"
checked. If this is the case, as long as the type isn't on outer edge of
the GroupItem, the GroupItem will nudge any PathItems that can be
aligned to the pixel grid onto it (crazy right?). Net net: Illustrator
will overlap a little bit of type to get the whole outline of the
GroupItem on the pixel grid.

Unfortunately, it does this AFTER we finish BlokContainer.layout(), so
we cache the wrong width. We don't have a great way of waiting for the
right layout to appear (it would be especially hard for nested
BlokContainers).

So the crappy fix is to check to see if layout rounding *could* be the
reason we report an invalid dimension in checkForRelayout(). This could
let a bug through in a very specific case, but that's why we use ===
instead of Utils.nearlyEqual(). I'd love to use PageItem.pixelAligned to
scope the hack even more, but that thing is right like 50% of the time
:(