Get TextFrames working better
Mainly for area text, which is a PITA because there aren't any APIs to
transform it's bounding box without scaling the damn type.

So first, any TextFrame that is AREATYPE is always drawn by Illustrator
according to it's geometric bounds. It's visible bounds exceed
Illustrator's controls. This is kind of handy, it lets us mess with the
type's bounding box so that we can do things like visually vertically
center it.

Since we're not using visual bounds everywhere, we can't always trust a
GroupItem when we ask for bounds, so there's some detection logic for
that.

In Blok.layout(), we don't have to deal with midpoint positioning
anymore, figured out how to scale from topleft instead of midpoint,
which makes the math easier. Also in layout(), we check to see whether
we're being asked to scale a TextFrame AREATEXT. If that's the case, we
have to do a nasty hack. Unfortunately, this is necessary because
resizing text containers is a core scenario (think responsive pages,
especially with flex-grow set).

Seeing as there are so many special cases, I added some more test cases
for text that reaffirm the common issues I was running into.

Note that when we scale a text area, the undo stack gets even more
wonky, but it doesn't end up being that bad of a UX.