Get rid of fixedWidth and fixedHeight, add <CTRL + R> shortcut
Holy cow this simplified things. Instead of trying to keep a fixed
width/height property for each Blok and allowing the user to modify
those sometimes (if we could detect it, which was basically only if you
had a Blok directly selected an did a transform), we just accept
whatever the current dims of any element are.

PROS:
- less confusing UX, I was finding that it's common to modify the size
of a Blok without ONLY it being directly selected. When your change gets
undone, it's very confusing
- way less code

CONS:
- resizing a BlokContainer resizes all children, which doesn't make the
most sense (conceptually, you shouldn't be able to resize it)
- You can't switch to stretch and back, you'll have to undo that change
- If you have space-between set, you can't resize the BlokContainer to
see your changes, you have to grab a Blok child, then move that, then
manually relayout

To help with manually relaying out (clicking the UI is a PITA), I sort
of hacked together a keyboard shortcut solution. We listen for the event
that fires before Illustrator's ruler is hidden or shown, which by
default is <CTRL + R> ("r" for "relayout", get it?) and do the
equivalent of clicking the "Layout" button. If you hit <CTRL + R> 2x in
under a second, we only listen to the first one. I assume this will be
somewhat common since you may not actually want to show or hide the
ruler.

Instead of <CTRL + R>, also looked at:

kAIDocumentTransparencyGridNotifier (CTRL + SHIFT + D)
kAIToolModifiersChangedNotifier - maybe just hitting <CTRL> could work?