Bug: coming back from Symbol Editing Mode takes a while
We were calling invalidate() blindly on every single Blok that changed
as a result of a Symbol change. This is fine if the Bloks are in
separate BlokContainer trees, but if they're in the same, it doesn't
help anything to call invalidate() more than once.

So we keep a list of the root BlokContainers and check that before
calling invalidate().