Add ease-of-use for space-between
We try to detect if a BlokContainer is the only thing selected and if
it's been resized and if it's set to space-between. In this one and only
case, we look at the cached dimensions and relayout, so as to actually
get spacing working.

Doesn't work if any type is part of the picture, it will scale. Not much
we can do about that, Illustrator's fault.