Bug: lazy Symbol creation sometimes didn't attach a Blok
1. Create a BlokContainer with normal PageItems inside
2. Turn one of those PageItems into a Symbol
3. Try to resize the Symbol or Blok (SymbolItem) that contains the
Symbol

ACTUAL: No relayout
EXPECTED: relayout

This happens because the PageItem that has the Blok tag attached to it
goes into the Symbol. If the BlokContainer did a true layout pass, then
the new SymbolItem child would have gotten a tag of its own, but since
no dimensions have changed, we can't trigger a layout, not even
explicitly.

The fix is simple, we change isBlokAttached()'s behavior to count
PageItems that don't have the tag, but have a direct parent that does
have the tag.