Bug: show/hide .spacer doesn't work on SymbolItems
If you name your Symbol with a ".spacer", every instance of it gets that
name by default (in the UI). But the actual SymbolItem.name on each is
blank. So if it's a SymbolItem with no name of its own, we can check
SymbolItem.symbol.name.

Also added more stringent logic so that we don't match ".spacers" or
"stuff.spacer". You must delimit with whitespace.