Improved Area Type handling
Made a big discovery that you CAN actually adjust Area Type's bounding
box via TextFrameItem.textPath. This let's us get rid of the nasty hack
in blok.ts's layout(), where we were creating a new Area Type and
copying content over (badly).

The rest of the updates are all to make resizing a BlokContainer in the
space-between scenario work. When the user resizes a GroupItem that
contains an AreaType, Illustrator "scales" the text in the Area Type by
adjusting some CharacterAttributes (horizontalScale, size, and leading).
horizontalScale is directly proportional to the width portion of the
scale. size is directly proportional to the height portion of the scale.
leading is adjusted at the same ratio as size ONLY if it isn't set to
Auto (CharacterAttributes.autoLeading = true). Oh and if it is
autoLeading = true, setting leading is harmless (ignored).

Anyway, to combat what Illustrator does, we cache those
CharacterAttribute values and reset them. Unfortunately, this creates a
bit of an awkward UX. When the user wants to update an Area Type that is
a Blok's character properties, they have to click a button in our UI to
update the cache of character properties. The alternative is that we
have to better detect the "resizing a BlokContainer" scenario. Maybe
that's worth the effort...

Fixed the BlokAdapter.isBlokAttached() behavior to be correct again,
instead adding a new method that more accurately expresses what we're
asking (should a Blok be attached?).