Even better Area Type handling
We were able to modify the previous approach through the
kAIArtPropertiesChangedNotifier, which is far more useful than
SelectionChanged since it fires when the font-size/leading/etc changes
on an Area Type, which gives us a chance to cache the new values. It
also fires when layers are reordered. It's a super noisy event though,
we must be careful.

Now that we can know exactly when Area Type is updated (must be
selected, must be a Blok. We won't know if someone is using character
styles though), we don't have to ask the user to manually update it.
Thus, the button is removed.

To be even more certain of the "resizing a BlokContainer" scenario, we
cache the previous selection. This further restricts it to direct
manipulation cases.