Bug: Area Type handling doesn't respect multiple font sizes in a single box
1) Make an Area Type with 2 lines of text, the first at 12px font size,
the second at 18px font size
2) add it to a distributed BlokContainer
3) Change the width of the BlokContainer

ACTUAL: text is not distorted, but all the sudden both the lines are
12px font size
EXPECTED: text is not distorted and the font size is preserved

We were trying to restore font size from cache, but we forgot that a
single Area Text can contain multiple ranges of text with different font
sizes. Rather than deal with the nightmare of caching all that data, we
use app.undo(), then reposition everything.

A side-effect of this is that you can't change the height of a
horizontal BlokContainer that's distributed. That's honestly OK.