UX improvements for resizing BlokContainers
Remember commit e08d36a? It turns out that there IS a way to not allow
the user to resize a vanilla BlokContainer WITHOUT having to keep track
of the cached width/height of everything.

It's actually quite simple. Instead of letting the user resize, then
running a layout pass to fix what they broke, we let the user resize,
record that size, then immediately undo the change! Then factor in the
"desired size" into our layout, but it's up to us what dimensions we
use.

We only use the a dimension on the primary axis if we're "distributed"
or if at least one child is flex-growing (this last part isn't
implemented yet).

We only use the dimension on the cross axis if we're "stretch" or if at
least one child is "stretch".

The second piece that makes this possible is having confidence that the
user is actually directly resizing a BlokContainer. If the user isn't
directly resizing, we use the current size of each PageItem and
GroupItem, no matter what.