Implement flex-grow
Bloks can now use flex-grow. Why do we need this? What if you want some
children to stay the same size, but others to increase in size?

Change comes with a test.

Fixed a bug in main.js: alignSelf wasn't always getting set correctly
because we spelled `viewModel` as `viewMode`.