More UI tweaks and added .bg padding regex
- got rid of Empty State, wasn't adding much. We solved it with a more
subtle prompt
- Divided the UI into 3 sections so that it doesn't jump around on you.
Blok settings will always appear in the same area (not always visible,
but you can move your mouse to the same spot every time). This is
similar to the Transform window. It has the same dividers and
hides/shows the Shape settings UI.
- Made the UI strings update depending on the
BlokContainer.flexDirection (or for the parent if its a Blok)
- better help tooltips
- copy tweaks
- Moved the relayout button to the bottom bar