Bug: couldn't toggle away from "Stretch"
If you set a BlokContainer.alignItems to STRETCH, then set it back to
FLEX-START, you visibly wouldn't see anything happen. This is because we
took the art's dims at face value. Sure, they're FLEX-START, but you
can't tell because they're all equal height.

Fix is just to have a separate cache for the stretch condition. We
restore it for the user.